*** empty log message ***
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@1048 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/focusc/xfocusc.cpp
@@ -21,6 +21,7 @@
 #include <pthread.h>
 #include <mcheck.h>
 #include <getopt.h>
+#include <math.h>
 
 #include "../utils/devcli.h"
 #include "../utils/devconn.h"
@@ -605,9 +606,9 @@ main (int argc, char **argv)
   // allocate colormap..
   for (int i = 0; i < 256; i++)
     {
-      rgb[i].red = 65536 * (1.0 * i / 256);
-      rgb[i].green = 65536 * (1.0 * i / 256);
-      rgb[i].blue = 65536 * (1.0 * i / 256);
+      rgb[i].red = (unsigned short) (65536 * (1.0 * i / 256));
+      rgb[i].green = (unsigned short) (65536 * (1.0 * i / 256));
+      rgb[i].blue = (unsigned short) (65536 * (1.0 * i / 256));
       rgb[i].flags = DoRed | DoGreen | DoBlue;
